SDC-1577. ConfigDef NUMBER should allow double type.
Reviewed-on: https://review.streamsets.net/1119

+++ common-ui/src/main/webapp/common/directives/codemirror/ui-codemirror.js
@@ -143,7 +143,7 @@ function uiCodemirrorDirective($timeout, uiCodemirrorConfig) {
         if(newValue === '') {
           newValue = 0;
         } else if(!isNaN(newValue)) {
-          newValue = parseInt(newValue);
+          newValue = parseFloat(newValue);
         }
       }
